backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
supabase==2.4.2
pydantic[email]==2.7.1
